replaced size_t with int

+++ include/string.h
@@ -1,10 +1,10 @@
 #ifndef STRING_H
 #define STRING_H
 
-size_t strlen(const char *str);
+int strlen(const char *str);
 char *strcpy(char *dest, const char *src);
-char *strncpy(char *dest, const char *src, size_t n);
+char *strncpy(char *dest, const char *src, int n);
 int strcmp(const char *str1, const char *str2);
-int strncmp(const char *str1, const char *str2, size_t n);
+int strncmp(const char *str1, const char *str2, int n);
 
 #endif
